fix up ManagePool to account for commission task
