+++ package.json
@@ -46,6 +46,7 @@
     "i18next": "^22.4.15",
     "i18next-browser-languagedetector": "^7.0.1",
     "lodash.throttle": "^4.1.1",
+    "rc-slider": "^10.1.1",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
